requirements.txt
asgiref==3.2.7
Django==3.0.5
django-countries==6.1.2
django-crispy-forms==1.9.0
django-sslserver==0.22
django-tempus-dominus==5.1.2.13
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.